backend/pyproject.toml
[tool.poetry]
name = "codeforge-backend"
version = "0.1.0"
description = "CodeForge AI Agent Engine - FastAPI backend with LLM orchestration"
authors = ["CodeForge Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.0.0"
langchain-anthropic = "^0.2.0"
langgraph = "^0.2.0"
supabase = "^2.8.0"
PyJWT = "^2.9.0"
cryptography = "^43.0.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
redis = "^5.1.0"
google-generativeai = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
black = "^24.8.0"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
